Backport 628e7702746ed4cc899d9727432b64c701533eb3
